+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/metrics/ApexMetricsComputer.java
@@ -4,7 +4,6 @@
 
 package net.sourceforge.pmd.lang.apex.metrics;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import net.sourceforge.pmd.annotation.InternalApi;
@@ -21,22 +20,14 @@ public class ApexMetricsComputer extends AbstractMetricsComputer<ASTUserClassOrI
 
     private static final ApexMetricsComputer INSTANCE = new ApexMetricsComputer();
 
+    @Override
+    protected List<ASTMethod> findOperations(ASTUserClassOrInterface<?> node) {
+        return ApexMetrics.findOps(node);
+    }
 
     @InternalApi
     public static ApexMetricsComputer getInstance() {
         return INSTANCE;
     }
 
-
-    @Override
-    protected List<ASTMethod> findOperations(ASTUserClassOrInterface<?> node) {
-        List<ASTMethod> candidates = node.findChildrenOfType(ASTMethod.class);
-        List<ASTMethod> result = new ArrayList<>(candidates);
-        for (ASTMethod method : candidates) {
-            if (method.getImage().matches("(<clinit>|<init>|clone)")) {
-                result.remove(method);
-            }
-        }
-        return result;
-    }
 }