+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/design/CyclomaticComplexityRule.java
@@ -16,6 +16,7 @@ import net.sourceforge.pmd.lang.apex.metrics.ApexMetrics;
 import net.sourceforge.pmd.lang.apex.metrics.api.ApexClassMetricKey;
 import net.sourceforge.pmd.lang.apex.metrics.api.ApexOperationMetricKey;
 import net.sourceforge.pmd.lang.apex.rule.AbstractApexRule;
+import net.sourceforge.pmd.lang.metrics.MetricsUtil;
 import net.sourceforge.pmd.lang.metrics.ResultOption;
 import net.sourceforge.pmd.properties.PropertyDescriptor;
 import net.sourceforge.pmd.properties.PropertyFactory;
@@ -23,7 +24,7 @@ import net.sourceforge.pmd.properties.PropertyFactory;
 
 /**
  * Cyclomatic complexity rule using metrics. Uses Wmc to report classes.
- * 
+ *
  * @author Clément Fournier
  */
 public class CyclomaticComplexityRule extends AbstractApexRule {
@@ -69,7 +70,7 @@ public class CyclomaticComplexityRule extends AbstractApexRule {
         classNames.pop();
 
         if (ApexClassMetricKey.WMC.supports(node)) {
-            int classWmc = (int) ApexMetrics.get(ApexClassMetricKey.WMC, node);
+            int classWmc = (int) MetricsUtil.computeMetric(ApexClassMetricKey.WMC, node);
 
             if (classWmc >= getProperty(CLASS_LEVEL_DESCRIPTOR)) {
                 int classHighest = (int) ApexMetrics.get(ApexOperationMetricKey.CYCLO, node, ResultOption.HIGHEST);
@@ -89,7 +90,7 @@ public class CyclomaticComplexityRule extends AbstractApexRule {
     @Override
     public final Object visit(ASTMethod node, Object data) {
 
-        int cyclo = (int) ApexMetrics.get(ApexOperationMetricKey.CYCLO, node);
+        int cyclo = (int) MetricsUtil.computeMetric(ApexOperationMetricKey.CYCLO, node);
         if (cyclo >= getProperty(METHOD_LEVEL_DESCRIPTOR)) {
             String opType = inTrigger ? "trigger"
                                       : node.getImage().equals(classNames.peek()) ? "constructor"