+++ pmd-core/src/main/java/net/sourceforge/pmd/processor/AbstractPMDProcessor.java
@@ -24,9 +24,6 @@ import net.sourceforge.pmd.SourceCodeProcessor;
 import net.sourceforge.pmd.benchmark.TimeTracker;
 import net.sourceforge.pmd.benchmark.TimedOperation;
 import net.sourceforge.pmd.benchmark.TimedOperationCategory;
-import net.sourceforge.pmd.lang.Language;
-import net.sourceforge.pmd.lang.LanguageRegistry;
-import net.sourceforge.pmd.lang.metrics.LanguageMetricsProvider;
 import net.sourceforge.pmd.renderers.Renderer;
 import net.sourceforge.pmd.util.datasource.DataSource;
 
@@ -37,7 +34,7 @@ import net.sourceforge.pmd.util.datasource.DataSource;
 public abstract class AbstractPMDProcessor {
 
     private static final Logger LOG = Logger.getLogger(AbstractPMDProcessor.class.getName());
-    
+
     protected final PMDConfiguration configuration;
 
     public AbstractPMDProcessor(PMDConfiguration configuration) {
@@ -56,7 +53,7 @@ public abstract class AbstractPMDProcessor {
     }
 
     /**
-     * 
+     *
      * @deprecated this method will be removed. It was once used to determine a short filename
      * for the file being analyzed, so that shortnames can be reported. But the logic has
      * been moved to the renderers.
@@ -78,15 +75,15 @@ public abstract class AbstractPMDProcessor {
      */
     protected RuleSets createRuleSets(RuleSetFactory factory, Report report) {
         final RuleSets rs = RulesetsFactoryUtils.getRuleSets(configuration.getRuleSets(), factory);
-        
+
         final Set<Rule> brokenRules = removeBrokenRules(rs);
         for (final Rule rule : brokenRules) {
             report.addConfigError(new Report.ConfigurationError(rule, rule.dysfunctionReason()));
         }
-        
+
         return rs;
     }
-    
+
     /**
      * Remove and return the misconfigured rules from the rulesets and log them
      * for good measure.
@@ -117,8 +114,6 @@ public abstract class AbstractPMDProcessor {
         configuration.getAnalysisCache().checkValidity(rs, configuration.getClassLoader());
         final SourceCodeProcessor processor = new SourceCodeProcessor(configuration);
 
-        resetMetrics();
-
         for (final DataSource dataSource : files) {
             // this is the real, canonical and absolute filename (not shortened)
             String realFileName = dataSource.getNiceFileName(false, null);
@@ -128,7 +123,7 @@ public abstract class AbstractPMDProcessor {
 
         // render base report first - general errors
         renderReports(renderers, ctx.getReport());
-        
+
         // then add analysis results per file
         collectReports(renderers);
 
@@ -136,15 +131,6 @@ public abstract class AbstractPMDProcessor {
         // the analysis is finished
         for (DataSource dataSource : files) {
             IOUtils.closeQuietly(dataSource);
-        }
-    }
-
-    private void resetMetrics() {
-        for (Language language : LanguageRegistry.getLanguages()) {
-            LanguageMetricsProvider<?, ?> languageMetricsProvider = language.getDefaultVersion().getLanguageVersionHandler().getLanguageMetricsProvider();
-            if (languageMetricsProvider != null) {
-                languageMetricsProvider.initialize();
-            }
         }
     }