+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/xpath/MetricFunction.java
@@ -19,9 +19,9 @@ import net.sourceforge.pmd.annotation.InternalApi;
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeDeclaration;
 import net.sourceforge.pmd.lang.java.ast.MethodLikeNode;
-import net.sourceforge.pmd.lang.java.metrics.JavaMetrics;
 import net.sourceforge.pmd.lang.java.metrics.api.JavaClassMetricKey;
 import net.sourceforge.pmd.lang.java.metrics.api.JavaOperationMetricKey;
+import net.sourceforge.pmd.lang.metrics.MetricsUtil;
 
 
 /**
@@ -82,9 +82,9 @@ public class MetricFunction implements Function {
 
     public static double getMetric(Node n, String metricKeyName) {
         if (n instanceof ASTAnyTypeDeclaration) {
-            return getClassMetric((ASTAnyTypeDeclaration) n, getClassMetricKey(metricKeyName));
+            return MetricsUtil.computeMetric(getClassMetricKey(metricKeyName), (ASTAnyTypeDeclaration) n);
         } else if (n instanceof MethodLikeNode) {
-            return getOpMetric((MethodLikeNode) n, getOperationMetricKey(metricKeyName));
+            return MetricsUtil.computeMetric(getOperationMetricKey(metricKeyName), (MethodLikeNode) n);
         } else {
             throw new IllegalStateException(genericBadNodeMessage());
         }
@@ -106,16 +106,6 @@ public class MetricFunction implements Function {
             throw new IllegalArgumentException(badOperationMetricKeyMessage());
         }
         return OPERATION_METRIC_KEY_MAP.get(constantName);
-    }
-
-
-    private static double getOpMetric(MethodLikeNode node, JavaOperationMetricKey key) {
-        return JavaMetrics.get(key, node);
-    }
-
-
-    private static double getClassMetric(ASTAnyTypeDeclaration node, JavaClassMetricKey key) {
-        return JavaMetrics.get(key, node);
     }